requirements.txt
streamlit
pandas
networkx
matplotlib.pyplot
numpy
random